refactor .py code for separation of concerns & OO abstractions
New abstractions added for IOC feed parsing, rule definition (and evaluation)
and the Event instances that are produced from evaluating rules.

I left RuleSet and Rule as partially defined, to keep things simple for now.
Event is basically schemaless as well, but the IOC feed parsing and how it
produces RuleSet instances is more well defined.  All of these are likely to
grow and evolve in upcoming commits and merges but this is a good place to
start.  It matches the features of the existing code and actually fixes a
couple of bugs.  Existing tests are refactored, more tests will be added soon.