distinguish IOC rule types from ruleset, maintain ruleset as a simple struct
The iocs/ directory is now ioc_formats to make it clearer that it is only dealing
with the file formats (and config formats) of various IOC sources.  The iocs.py
module has been hoisted up to main, alongside event.py (to be populated more
explicitly in a follow-up when the Event schema has stabilized) and all rules-
related code is now in a matchers/ directory, to separate monitor (observation)
production and the logic needed for rule matching.