refactor .py code for separation of concerns & OO abstractions (#8)
* refactor .py code for separation of concerns & OO abstractions

New abstractions added for IOC feed parsing, rule definition (and evaluation)
and the Event instances that are produced from evaluating rules.

I left RuleSet and Rule as partially defined, to keep things simple for now.
Event is basically schemaless as well, but the IOC feed parsing and how it
produces RuleSet instances is more well defined.  All of these are likely to
grow and evolve in upcoming commits and merges but this is a good place to
start.  It matches the features of the existing code and actually fixes a
couple of bugs.  Existing tests are refactored, more tests will be added soon.

* add tests for the rules module (add rule, merge rules, process endpoint)

* update pydoc for rules.py

Co-authored-by: Vinicius Fortuna <fortuna@users.noreply.github.com>

* addresses review comments, mostly renaming, some restructuring

* rename *Scanner to *Monitor

* further refactoring of the monitor classes and related code.

also adds a process executor pool and runs the tshark monitors within that,
it will help propagate exceptions up to the main process.

* distinguish IOC rule types from ruleset, maintain ruleset as a simple struct

The iocs/ directory is now ioc_formats to make it clearer that it is only dealing
with the file formats (and config formats) of various IOC sources.  The iocs.py
module has been hoisted up to main, alongside event.py (to be populated more
explicitly in a follow-up when the Event schema has stabilized) and all rules-
related code is now in a matchers/ directory, to separate monitor (observation)
production and the logic needed for rule matching.

* add tests for IP matching and TLS connection matching, some typing

added types to constructors and methods that are called from other
modules

* some finishing cleanup, use datetime for timestamps and small naming change